Debug some sample

+++ webcl/hello_world_2/hello_world_2.cl
@@ -1,7 +1,7 @@
 
-__kernel void hello_kernel(__global const float *a,
-						__global const float *b,
-						__global float *result)
+__kernel void hello_kernel(__global const int *a,
+						__global const int *b,
+						__global int *result)
 {
     int gid = get_global_id(0);
